requirements.txt
playwright==1.52.0
playwright-stealth==2.0.3
pyyaml==6.0.2
requests==2.32.3
pytesseract==0.3.13
Pillow==11.2.1
